feat(29-03): migrate builder from FNV-1a to nt_hash module
- Delete nt_builder_fnv1a() and FNV-1a constants from nt_builder_hash.c
- Replace nt_builder_hash() with nt_builder_normalize_and_hash() (64-bit)
- Migrate all call sites in nt_builder.c to nt_hash64_str/nt_hash32_str
- Attribute name hashing in nt_builder_mesh.c uses nt_hash32_str
- Link nt_hash in builder CMakeLists.txt
- Update import function signatures to uint64_t resource_id
- Update NtBuildEntry.resource_id to uint64_t
- Update dump utility for 64-bit resource_id display
- Migrate all test_builder tests to nt_builder_normalize_and_hash

+++ tools/builder/nt_builder_shader.c
@@ -103,7 +103,7 @@ static void collapse_whitespace(char *buf, uint32_t *len) {
 /* --- Shader import (called from finish_pack) --- */
 
 // NOLINTNEXTLINE(readability-function-cognitive-complexity)
-nt_build_result_t nt_builder_import_shader(NtBuilderContext *ctx, const char *path, nt_build_shader_stage_t stage, uint32_t resource_id) {
+nt_build_result_t nt_builder_import_shader(NtBuilderContext *ctx, const char *path, nt_build_shader_stage_t stage, uint64_t resource_id) {
     if (!ctx || !path) {
         return NT_BUILD_ERR_VALIDATION;
     }
